refactor: layout folder structure
